requirements.txt
Django==1.11
django-extensions==1.9.7
django-debug-toolbar==1.6
django-filter==1.1.0
djangorestframework==3.6.0
djangorestframework-filters==0.10.2
Markdown==2.1.0
